chore(theme): Move defaults to separate files
