+++ lib/theme/breakpoints.js
@@ -1,0 +1,8 @@
+const breakpoints = [
+  '40em', //  640px - mobile
+  '52em', //  832px - tablet
+  '64em', // 1024px - desktop
+  '88em', // 1408px - widescreen
+];
+
+export default breakpoints;
