+++ lib/theme/index.js
@@ -1,137 +1,27 @@
 import { variant } from 'styled-system';
 import { transparentize } from 'polished';
 
-const ocBrandColors = {
-  900: '#0C2D66',
-  800: '#1041A3',
-  700: '#1153D6',
-  600: '#1869F5',
-  500: '#297EFF',
-  400: '#5CA3FF',
-  300: '#8FC7FF',
-  200: '#C2E2FF',
-  100: '#E6F3FF',
-  50: '#F5FAFF',
-};
+// Import defaults
+import defaultBreakpoints from './breakpoints';
+import defaultColors from './colors';
+import defaultSpaces from './spaces';
+import { defaultFontSizes, defaultLineHeights } from './text-sizes';
 
-const defaultColors = {
-  /** White */
-  white: {
-    full: '#FFFFFF',
-    transparent: {
-      72: 'rgba(255, 255, 255, 0.72)',
-      48: 'rgba(255, 255, 255, 0.48)',
-    },
-  },
-  /** Black, neutral */
-  black: {
-    900: '#141414',
-    800: '#313233',
-    700: '#4E5052',
-    600: '#76777A',
-    500: '#9D9FA3',
-    400: '#C4C7CC',
-    300: '#DCDEE0',
-    200: '#E8E9EB',
-    100: '#F0F1F2',
-    50: '#F7F8FA',
-    transparent: {
-      90: 'rgba(19, 20, 20, 0.90)',
-      80: 'rgba(19, 20, 20, 0.80)',
-      40: 'rgba(19, 20, 20, 0.4)',
-      20: 'rgba(19, 20, 20, 0.2)',
-      8: 'rgba(19, 20, 20, 0.08)',
-    },
-  },
-  /** Primary */
-  primary: ocBrandColors,
-  /** Info - right now the colors are the same than prumary */
-  blue: ocBrandColors,
-  /** Success, Approve */
-  green: {
-    900: '#005728',
-    800: '#007A39',
-    700: '#009E4A',
-    600: '#16B861',
-    500: '#18CC6C',
-    400: '#36E085',
-    300: '#83EBB4',
-    200: '#BEFADA',
-    100: '#D6FFE9',
-    50: '#F0FFF7',
-  },
-  /** Warning, Alert */
-  yellow: {
-    900: '#703C00',
-    800: '#945400',
-    700: '#B86E00',
-    600: '#DB8B00',
-    500: '#FFAA00',
-    400: '#FFC94D',
-    300: '#FFEB85',
-    200: '#FFF5C2',
-    100: '#FFFAE0',
-  },
-  /** Danger, Error */
-  red: {
-    900: '#5C071E',
-    800: '#850B2B',
-    700: '#B80F3C',
-    600: '#D60940',
-    500: '#F51D57',
-    400: '#FF4778',
-    300: '#FF85A5',
-    200: '#FFC2D2',
-    100: '#FFEBF0',
-    50: '#FFF7F9',
-  },
-};
-
-const defaultFontSizes = {
-  H1: 52,
-  H2: 40,
-  H3: 32,
-  H4: 24,
-  H5: 20,
-  H6: 9,
-  LeadParagraph: 16,
-  Paragraph: 14,
-  Caption: 12,
-  Tiny: 10,
-};
-
-const defaultLineHeights = {
-  H1: '56px',
-  H2: '44px',
-  H3: '36px',
-  H4: '32px',
-  H5: '24px',
-  H6: '14px',
-  LeadParagraph: '24px',
-  Paragraph: '20px',
-  Caption: '18px',
-  Tiny: '14px',
-};
-
-// using default space values from styled-system
-const defaultSpace = [0, 4, 8, 16, 32, 64, 128, 256, 512];
+function toPx(value) {
+  return `${value}px`;
+}
 
 export const generateTheme = ({
   colors = defaultColors,
   fontSizes = defaultFontSizes,
   lineHeights = defaultLineHeights,
-  space = defaultSpace,
+  space = defaultSpaces,
 } = {}) => ({
   colors,
   fontSizes,
   lineHeights,
   space,
-  breakpoints: [
-    '40em', //  640px - mobile
-    '52em', //  832px - tablet
-    '64em', // 1024px - desktop
-    '88em', // 1408px - widescreen
-  ],
+  breakpoints: defaultBreakpoints,
   buttons: {
     standard: {
       backgroundColor: 'white',
@@ -267,8 +157,7 @@ export const generateTheme = ({
   },
 });
 
-const theme = generateTheme();
-
+// Export variants
 export const buttonStyle = variant({
   key: 'buttons',
   prop: 'buttonStyle',
@@ -284,8 +173,6 @@ export const messageType = variant({
   prop: 'type',
 });
 
+// Generate and export main theme
+const theme = generateTheme();
 export default theme;
-
-function toPx(value) {
-  return `${value}px`;
-}
