+++ lib/theme/spaces.js
@@ -1,0 +1,3 @@
+// using default space values from styled-system
+const defaultSpaces = [0, 4, 8, 16, 32, 64, 128, 256, 512];
+export default defaultSpaces;
