+++ lib/theme/text-sizes.js
@@ -1,0 +1,25 @@
+export const defaultFontSizes = {
+  H1: 52,
+  H2: 40,
+  H3: 32,
+  H4: 24,
+  H5: 20,
+  H6: 9,
+  LeadParagraph: 16,
+  Paragraph: 14,
+  Caption: 12,
+  Tiny: 10,
+};
+
+export const defaultLineHeights = {
+  H1: '56px',
+  H2: '44px',
+  H3: '36px',
+  H4: '32px',
+  H5: '24px',
+  H6: '14px',
+  LeadParagraph: '24px',
+  Paragraph: '20px',
+  Caption: '18px',
+  Tiny: '14px',
+};
